Update: rewrote all remaining resource classes
