+++ src/api/search/index.js
@@ -1,65 +1,15 @@
 // @flow
 import Resource from '../resource';
 import type { APIPromise } from '../resource';
-
-type SearchDocumentsQuery = {
-    q?: string
-}
-
-type SearchDocumentsParams = {
-    query?: SearchDocumentsQuery
-}
-
-type SearchOrganizationQuery = {
-    q?: string,
-    limit?: number,
-    offset?: number,
-    certification_id?: number
-}
-
-type SearchOrganizationParams = {
-    organization_id: number,
-    index_type: string,
-    query?: SearchOrganizationQuery
-}
-
-type searchResultsSchema = {
-    took: number,
-    timed_out: boolean,
-    _shards: {
-        successful: number,
-        failed: number,
-        total: number
-    },
-    hits: {
-        total: number,
-        max_score: number,
-        hits: Array<{
-            _score: number,
-            _id: number,
-            _type: string,
-            _index: string,
-            _source: Object,
-            highlight?: Object
-        }>
-    }
-}
-
-type SearchDocumentsData = [
-    searchResultsSchema
-]
-
-type CreateSearchData = [
-    string
-]
-
-type SearchOrganizationData = [
-    searchResultsSchema
-]
+import type {
+    ESSearch,
+    SearchDocumentsParams,
+    SearchOrganizationParams
+} from './types';
 
 export default class Search extends Resource {
     /**
-     * @api {get} /v1/search
+     * @api {get} /v1/search searchDocuments
      * @apiGroup Search
      * @apiName searchDocuments
      * @apiDescription Search all ES docs for the given query_string. Returns results in ES idiom.
@@ -67,37 +17,36 @@ export default class Search extends Resource {
      * @apiExample {js} Example:
      *             gigwalk.customers.searchDocuments({...})
      */
-    searchDocuments(params: SearchDocumentsParams): APIPromise<SearchDocumentsData> {
+    searchDocuments(params: SearchDocumentsParams): APIPromise<[ESSearch]> {
         const queryString = this.queryStringForSearchObject(params.query);
 
         return this.client.get(`/v1/search${queryString}`);
     }
 
     /**
-     * @api {post} /v1/search
+     * @api {post} /v1/search create
      * @apiGroup Search
-     * @apiName createSearch
+     * @apiName create
      * @apiDescription Not implemented.
      * @apiExample {js} Example:
-     *             gigwalk.search.createSearch({...})
+     *             gigwalk.search.create({...})
      */
-    createSearch(): APIPromise<CreateSearchData> {
+    create(): APIPromise<[string]> {
         return this.client.post('/v1/search');
     }
 
     /**
-     * @api {get} /v1/organizations/{organization_id}/search/{index_type}
+     * @api {get} /v1/organizations/:organization_id/search/:index_type searchOrganization
      * @apiGroup Search
      * @apiName searchOrganization
-     * @apiDescription Search in an organization within groups, members, location_lists, target_lists, tickets or subscriptions.
-                       Capable of returning paginated results.
+     * @apiDescription Search in an organization. Capable of returning paginated results.
      * @apiParam {Number} organization_id
      * @apiParam {String} index_type
      * @apiParam {Object} query
      * @apiExample {js} Example:
      *             gigwalk.search.searchOrganization({...})
      */
-    searchOrganization(params: SearchOrganizationParams): APIPromise<SearchOrganizationData> {
+    searchOrganization(params: SearchOrganizationParams): APIPromise<[ESSearch]> {
         const queryString = this.queryStringForSearchObject(params.query);
 
         return this.client.get(`/v1/organizations/${params.organization_id}/search/${params.index_type}${queryString}`);
