+++ src/api/targetLists/index.js
@@ -1,326 +1,180 @@
 // @flow
 import Resource from '../resource';
 import type { APIPromise } from '../resource';
-
-type TargetTemplate = {
-    title: string
-}
-
-type TargetListTemplate = {
-  name: string,
-  observation_target_type: number,
-  observation_targets: Array<TargetTemplate>,
-  status?: string
-}
-
-type GetTargetListsQuery = {
-    limit?: number,
-    offset?: number,
-    order_by?: string,
-    order_dir?: 'ASCENDING' | 'DESCENDING',
-    include_auto_generated?: number
-}
-
-type GetTargetListsParams = {
-    query?: GetTargetListsQuery
-}
-
-type GetTargetListParams = {
-    observation_target_list_id: number
-}
-
-type GetOrganizationTargetListsQuery = {
-    limit?: number,
-    offset?: number,
-    order_by?: string,
-    order_dir?: 'ASCENDING' | 'DESCENDING',
-    auto_generated?: number
-}
-
-type GetOrganizationTargetListsParams = {
-    organization_id: number,
-    query?: GetOrganizationTargetListsQuery
-}
-
-type CreateOrganizationTargetListParams = {
-    organization_id: number,
-    target_list: TargetListTemplate
-}
-
-type GetOrganzationTargetListParams = {
-    organization_id: number,
-    observation_target_list_id: number
-}
-
-type DeleteTargetListParams = {
-    observation_target_list_id: number
-}
-
-type UpdateTargetListParams = {
-    observation_target_list_id: number,
-    target_list: TargetListTemplate
-}
-
-type SearchTargetsInObservationListQuery = {
-    q?: string,
-    limit?: number,
-    offset?: number,
-    order_by?: string,
-    order_dir?: 'ASCENDING' | 'DESCENDING'
-}
-
-type SearchTargetsInObservationListParams = {
-    target_list_id: number,
-    query?: SearchTargetsInObservationListQuery
-}
-
-type SearchTargetsInListQuery = {
-    q?: string,
-    limit?: number,
-    offset?: number,
-    order_by?: string,
-    order_dir?: 'ASCENDING' | 'DESCENDING'
-}
-
-type SearchTargetsInListParams = {
-    target_list_id: number,
-    query?: SearchTargetsInListQuery
-}
-
-type GetTargetsFromListQuery = {
-    limit?: number,
-    offset?: number,
-    order_by?: string,
-    order_dir?: 'ASCENDING' | 'DESCENDING'
-}
-
-type GetTargetsFromListParams = {
-    target_list_id: number,
-    query?: GetTargetsFromListQuery
-}
-
-type UpdateTargetsInListParams = {
-    target_list_id: number,
-    action: string,
-    target_ids: Array<number>
-}
-
-type SearchDataItemsInListParams = {
-    observation_target_id: number,
-    location_id: number,
-    item_count: number
-}
-
-type TargetListSummarySchema = {
-    name: string,
-    id: number,
-    observation_target_count: number,
-    observation_target_type_id: number,
-    status: string
-}
-
-type ObservationTargetSchema = {
-    title: string,
-    id: number,
-    organization_id: number,
-    observation_target_type_id: number
-}
-
-type GetTargetListsData = Array<TargetListSummarySchema>
-
-type GetTargetListData = [
-    TargetListSummarySchema
-]
-
-type GetOrganizationTargetListsData = Array<TargetListSummarySchema>
-
-type CreateOrganizationTargetListData = [
-    TargetListSummarySchema
-]
-
-type GetOrganzationTargetListData = [
-    TargetListSummarySchema
-]
-
-type DeleteTargetListData = [
-    number
-]
-
-type UpdateTargetListData = [
-    TargetListSummarySchema
-]
-
-type SearchTargetsInObservationListData = [ // NEED TO CHECK
-
-]
-
-type SearchTargetsInListData = [ // NEED TO CHECK
-
-]
-
-type GetTargetsFromListData = Array<ObservationTargetSchema>
-
-type UpdateTargetsInListData = Array<ObservationTargetSchema>
-
-type SearchDataItemsInListData = Array<ObservationTargetSchema> // NEED TO CHECK
+import type {
+    TargetList,
+    ObservationTarget,
+    GetTargetListsParams,
+    GetTargetListParams,
+    GetOrganizationTargetListsParams,
+    CreateOrganizationTargetListParams,
+    GetOrganizationTargetListParams,
+    DeleteTargetListParams,
+    UpdateTargetListParams,
+    SearchTargetsInObservationListParams,
+    SearchTargetsInListParams,
+    GetTargetsFromListParams,
+    UpdateTargetsInListParams,
+    SearchDataItemsInListParams
+} from './types';
 
 export default class TargetLists extends Resource {
     /**
-     * @api {get} /v1/organization_observation_target_lists
+     * @api {get} /v1/organization_observation_target_lists getAll
      * @apiGroup TargetLists
-     * @apiName getTargetLists
+     * @apiName getAll
      * @apiDescription Return all target lists. This can only be invoked by the platform admin. Capable of returning paginated results.
      * @apiParam {Object} query
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.getTargetLists({...})
+     *             gigwalk.targetLists.getAll({...})
      */
-    getTargetLists(params: GetTargetListsParams): APIPromise<GetTargetListsData> {
+    getAll(params: GetTargetListsParams): APIPromise<Array<TargetList>> {
         const queryString = (params) ? this.queryStringForSearchObject(params.query) : '';
 
         return this.client.get(`/v1/organization_observation_target_lists${queryString}`);
     }
 
     /**
-     * @api {get} /v1/organization_observation_target_lists/{observation_target_list_id}
+     * @api {get} /v1/organization_observation_target_lists/:observation_target_list_id get
      * @apiGroup TargetLists
-     * @apiName getTargetList
+     * @apiName get
      * @apiDescription Get org_observation_target_list.
      * @apiParam {Number} observation_target_list_id
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.getTargetList({...})
+     *             gigwalk.targetLists.get({...})
      */
-    getTargetList(params: GetTargetListParams): APIPromise<GetTargetListData> {
+    get(params: GetTargetListParams): APIPromise<[TargetList]> {
         return this.client.get(`/v1/organization_observation_target_lists/${params.observation_target_list_id}`);
     }
 
     /**
-     * @api {get} /v1/organizations/{organization_id}/target_lists
+     * @api {get} /v1/organizations/:organization_id/target_lists getAllForOrganization
      * @apiGroup TargetLists
-     * @apiName getOrganizationTargetLists
+     * @apiName getAllForOrganization
      * @apiDescription Get all organization_observation_target_lists in organiation. Capable of returning paginated results.
      * @apiParam {Number} organization_id
      * @apiParam {Object} query
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.getOrganizationTargetLists({...})
+     *             gigwalk.targetLists.getAllForOrganization({...})
      */
-    getOrganizationTargetLists(params: GetOrganizationTargetListsParams): APIPromise<GetOrganizationTargetListsData> {
+    getAllForOrganization(params: GetOrganizationTargetListsParams): APIPromise<Array<TargetList>> {
         const queryString = this.queryStringForSearchObject(params.query);
 
         return this.client.get(`/v1/organizations/${params.organization_id}/target_lists${queryString}`);
     }
 
     /**
-     * @api {post} /v1/organizations/{organization_id}/target_lists
+     * @api {post} /v1/organizations/:organization_id/target_lists createForOrganization
      * @apiGroup TargetLists
-     * @apiName createOrganizationTargetList
+     * @apiName createForOrganization
      * @apiDescription Create a new org_observation_target_list. Currently, there are no permissions checks.
      * @apiParam {Number} organization_id
      * @apiParam {Object} target_list
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.createOrganizationTargetList({...})
+     *             gigwalk.targetLists.createForOrganization({...})
      */
-    createOrganizationTargetList(params: CreateOrganizationTargetListParams): APIPromise<CreateOrganizationTargetListData> {
+    createForOrganization(params: CreateOrganizationTargetListParams): APIPromise<[TargetList]> {
         return this.client.post(`/v1/organizations/${params.organization_id}/target_lists`, { ...params.target_list });
     }
 
     /**
-     * @api {get} /v1/organizations/{organization_id}/target_lists/{observation_target_list_id}
+     * @api {get} /v1/organizations/:organization_id/target_lists/:observation_target_list_id getForOrganization
      * @apiGroup TargetLists
-     * @apiName getOrganzationTargetList
+     * @apiName getForOrganization
      * @apiDescription Get organization_observation_target_list.
      * @apiParam {Number} organization_id
      * @apiParam {Number} observation_target_list_id
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.getOrganzationTargetList({...})
+     *             gigwalk.targetLists.getForOrganization({...})
      */
-    getOrganzationTargetList(params: GetOrganzationTargetListParams): APIPromise<GetOrganzationTargetListData> {
+    getForOrganization(params: GetOrganizationTargetListParams): APIPromise<[TargetList]> {
         return this.client.get(`/v1/organizations/${params.organization_id}/target_lists/${params.observation_target_list_id}`);
     }
 
     /**
-     * @api {delete} /v1/organization_observation_target_lists/{observation_target_list_id}
+     * @api {delete} /v1/organization_observation_target_lists/:observation_target_list_id delete
      * @apiGroup TargetLists
-     * @apiName deleteTargetList
+     * @apiName delete
      * @apiDescription Delete org_obs_target_list_id. This is a soft delete; org_obs_target_list marked as DELETED.
      * @apiParam {Number} observation_target_list_id
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.deleteTargetList({...})
+     *             gigwalk.targetLists.delete({...})
      */
-    deleteTargetList(params: DeleteTargetListParams): APIPromise<DeleteTargetListData> {
+    delete(params: DeleteTargetListParams): APIPromise<[number]> {
         return this.client.delete(`/v1/organization_observation_target_lists/${params.observation_target_list_id}`);
     }
 
     /**
-     * @api {put} /v1/organization_observation_target_lists/{observation_target_list_id}
+     * @api {put} /v1/organization_observation_target_lists/:observation_target_list_id update
      * @apiGroup TargetLists
-     * @apiName updateTargetList
+     * @apiName update
      * @apiDescription Modify the specified org_obs_target_list.
      * @apiParam {Number} observation_target_list_id
      * @apiParam {Object} target_list
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.updateTargetList({...})
+     *             gigwalk.targetLists.update({...})
      */
-    updateTargetList(params: UpdateTargetListParams): APIPromise<UpdateTargetListData> {
+    update(params: UpdateTargetListParams): APIPromise<[TargetList]> {
         return this.client.put(`/v1/organization_observation_target_lists/${params.observation_target_list_id}`, { ...params.target_list });
     }
 
     /**
-     * @api {post} /v1/organization_observation_target_lists/{target_list_id}/search/observation_targets
+     * @api {post} /v1/organization_observation_target_lists/:target_list_id/search/observation_targets searchObservationList
      * @apiGroup TargetLists
-     * @apiName searchTargetsInObservationList
+     * @apiName searchObservationList
      * @apiDescription Search all targets of an org_obs_target_list for the specified query_string. Capable of returning paginated results.
      * @apiParam {Number} target_list_id
      * @apiParam {Object} query
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.searchTargetsInObservationList({...})
+     *             gigwalk.targetLists.searchObservationList({...})
      */
-    searchTargetsInObservationList(params: SearchTargetsInObservationListParams): APIPromise<SearchTargetsInObservationListData> {
+    searchObservationList(params: SearchTargetsInObservationListParams): APIPromise<[]> {
         const queryString = this.queryStringForSearchObject(params.query);
 
         return this.client.post(`/v1/organization_observation_target_lists/${params.target_list_id}/search/observation_targets${queryString}`);
     }
 
     /**
-     * @api {post} /v1/target_lists/{target_list_id}/search/targets
+     * @api {post} /v1/target_lists/:target_list_id/search/targets searchList
      * @apiGroup TargetLists
-     * @apiName searchTargetsInList
+     * @apiName searchList
      * @apiDescription Search all targets of an org_obs_target_list for the specified query_string. Capable of returning paginated results.
      * @apiParam {Number} target_list_id
      * @apiParam {Object} query
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.searchTargetsInList({...})
+     *             gigwalk.targetLists.searchList({...})
      */
-    searchTargetsInList(params: SearchTargetsInListParams): APIPromise<SearchTargetsInListData> {
+    searchList(params: SearchTargetsInListParams): APIPromise<[]> {
         const queryString = this.queryStringForSearchObject(params.query);
 
         return this.client.post(`/v1/target_lists/${params.target_list_id}/search/targets${queryString}`);
     }
 
     /**
-     * @api {get} /v1/organization_observation_target_lists/{target_list_id}/observation_targets
+     * @api {get} /v1/organization_observation_target_lists/:target_list_id/observation_targets getTargets
      * @apiGroup TargetLists
-     * @apiName getTargetsFromList
+     * @apiName getTargets
      * @apiDescription Return all targets of the specified org_obs_target_list.
      * @apiParam {Number} target_list_id
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.getTargetsFromList({...})
+     *             gigwalk.targetLists.getTargets({...})
      */
-    getTargetsFromList(params: GetTargetsFromListParams): APIPromise<GetTargetsFromListData> {
+    getTargets(params: GetTargetsFromListParams): APIPromise<Array<ObservationTarget>> {
         return this.client.get(`/v1/organization_observation_target_lists/${params.target_list_id}/observation_targets`);
     }
 
     /**
-     * @api {put} /v1/organization_observation_target_lists/{target_list_id}/observation_targets
+     * @api {put} /v1/organization_observation_target_lists/:target_list_id/observation_targets updateTargets
      * @apiGroup TargetLists
-     * @apiName updateTargetsInList
+     * @apiName updateTargets
      * @apiDescription ADD or REMOVE targets from the specified org_obs_target_list.
      * @apiParam {Number} target_list_id
      * @apiParam {String} action
      * @apiParam {Array<Number>} target_ids
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.updateTargetsInList({...})
+     *             gigwalk.targetLists.updateTargets({...})
      */
-    updateTargetsInList(params: UpdateTargetsInListParams): APIPromise<UpdateTargetsInListData> {
+    updateTargets(params: UpdateTargetsInListParams): APIPromise<Array<ObservationTarget>> {
         const data = {
             action: params.action,
             target_ids: params.target_ids
@@ -330,18 +184,18 @@ export default class TargetLists extends Resource {
     }
 
     /**
-     * @api {post} /v1/target_lists/target_history
+     * @api {post} /v1/target_lists/target_history searchDataItems
      * @apiGroup TargetLists
-     * @apiName searchDataItemsInList
+     * @apiName searchDataItems
      * @apiDescription Search all the data_items for the specified obs_target_id and location_id. Results sorted by data_item_timestamp.
                        Photo data_items will be excluded from returned data.
      * @apiParam {Number} observation_target_id
      * @apiParam {Number} location_id
      * @apiParam {Number} item_count
      * @apiExample {js} Example:
-     *             gigwalk.targetLists.searchDataItemsInList({...})
+     *             gigwalk.targetLists.searchDataItems({...})
      */
-    searchDataItemsInList(params: SearchDataItemsInListParams): APIPromise<SearchDataItemsInListData> {
+    searchDataItems(params: SearchDataItemsInListParams): APIPromise<Array<ObservationTarget>> {
         const data = {
             observation_target_id: params.observation_target_id,
             location_id: params.location_id,
